refactor(modules): Refactor module names and imports
* Rename main entrypoint
* Sort imports
* Rename module names
